+++ Tests/SwiftyPyStringTests/FormatTests.swift
@@ -43,40 +43,6 @@ final class FormatTests : XCTestCase {
     func testFormatPositional() throws {
         XCTAssertEqual("{0} # {1} # {0}".format("@","&"), "@ # & # @")
     }
-    func testStringFormatAlign() throws {
-        XCTAssertEqual("{:5}".format("s"), "s    ")
-        XCTAssertEqual("{:<5}".format("s"), "s    ")
-        XCTAssertEqual("{:^5}".format("s"), "  s  ")
-        XCTAssertEqual("{:>5}".format("s"), "    s")
-    }
-    func testIntegerFormatAlign() throws {
-        XCTAssertEqual("{:5}".format(1), "    1")
-        XCTAssertEqual("{:<5}".format(1), "1    ")
-        XCTAssertEqual("{:^5}".format(1), "  1  ")
-        XCTAssertEqual("{:>5}".format(1), "    1")
-    }
-    func testFloatFormatAlign() throws {
-        XCTAssertEqual("{:5}".format(1.1), "  1.1")
-        XCTAssertEqual("{:<5}".format(1.1), "1.1  ")
-        XCTAssertEqual("{:^5}".format(1.1), " 1.1 ")
-        XCTAssertEqual("{:>5}".format(1.1), "  1.1")
-        XCTAssertEqual("{:5}".format(1.0), "  1.0")
-    }
-    func testStringFormatFill() throws {
-        XCTAssertEqual("{:0<5}".format("s"), "s0000")
-        XCTAssertEqual("{:0^5}".format("s"), "00s00")
-        XCTAssertEqual("{:0>5}".format("s"), "0000s")
-    }
-    func testIntegerFormatFill() throws {
-        XCTAssertEqual("{:a<5}".format(1), "1aaaa")
-        XCTAssertEqual("{:a^5}".format(1), "aa1aa")
-        XCTAssertEqual("{:a>5}".format(1), "aaaa1")
-    }
-    func testFloatFormatFill() throws {
-        XCTAssertEqual("{:a<5}".format(1.1), "1.1aa")
-        XCTAssertEqual("{:a^5}".format(1.1), "a1.1a")
-        XCTAssertEqual("{:a>5}".format(1.1), "aa1.1")
-    }
     /// Returns path to the built products directory.
     var productsDirectory: URL {
         #if os(macOS)
